+++ packages/expo-modules-core/android/src/compose/expo/modules/kotlin/views/ExpoComposeView.kt
@@ -56,13 +56,6 @@ abstract class ExpoComposeView<T : ComposeProps>(
 
   @Composable
   protected fun Children(composableScope: ComposableScope) {
-    if (withHostingView) {
-      with(composableScope) {
-        Content()
-      }
-      return
-    }
-
     for (index in 0..<this.size) {
       val child = getChildAt(index) as? ExpoComposeView<*> ?: continue
       with(composableScope) {
@@ -72,6 +65,7 @@ abstract class ExpoComposeView<T : ComposeProps>(
       }
     }
   }
+    
 
   init {
     if (withHostingView) {
@@ -87,7 +81,7 @@ abstract class ExpoComposeView<T : ComposeProps>(
       it.layoutParams = LayoutParams(LayoutParams.MATCH_PARENT, LayoutParams.MATCH_PARENT)
       it.setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
       it.setContent {
-        Children(ComposableScope())
+        Content()
       }
       it.addOnAttachStateChangeListener(object : OnAttachStateChangeListener {
         override fun onViewAttachedToWindow(v: View) {