+++ packages/expo-ui/android/src/main/java/expo/modules/ui/Modifiers.kt
@@ -20,13 +20,13 @@ fun Modifier.fromExpoModifiers(
   modifiers: List<ExpoModifier>,
   composableScope: ComposableScope? = null
 ): Modifier {
-  return modifiers.fold(this) { acc, modifier ->
+  return modifiers?.fold(this) { acc, modifier ->
     composableScope?.let { holder ->
       modifier.getFromScope?.invoke(holder)?.let {
-        return@fold acc.then(it)
+        return@fold it.then(acc)
       }
     }
-    modifier.ref?.let { acc.then(it) } ?: acc
+    modifier.ref?.let { it.then(acc) } ?: acc
   }
 }