+++ packages/expo-ui/src/jetpack-compose/Shape/index.tsx
@@ -42,23 +42,63 @@ const ShapeNativeView: React.ComponentType<any> = requireNativeView('ExpoUI', 'S
 export type ShapeJSXElement = React.JSX.Element & { __expo_shape_jsx_element_marker: true };
 
 function Star(props: ShapeProps) {
-  return (<ShapeNativeView {...props} style={props?.style} type="star" />) as ShapeJSXElement;
+  return (
+    <ShapeNativeView
+      {...props}
+      style={props?.style}
+      // @ts-expect-error
+      modifiers={props?.modifiers?.map((m) => m.__expo_shared_object_id__)}
+      type="star"
+    />
+  ) as ShapeJSXElement;
 }
 
 function PillStar(props: ShapeProps) {
-  return (<ShapeNativeView {...props} style={props?.style} type="pillStar" />) as ShapeJSXElement;
+  return (
+    <ShapeNativeView
+      {...props}
+      style={props?.style}
+      // @ts-expect-error
+      modifiers={props?.modifiers?.map((m) => m.__expo_shared_object_id__)}
+      type="pillStar"
+    />
+  ) as ShapeJSXElement;
 }
 
 function Pill(props: Pick<ShapeProps, 'smoothing' | 'style' | 'color'>) {
-  return (<ShapeNativeView {...props} style={props?.style} type="pill" />) as ShapeJSXElement;
+  return (
+    <ShapeNativeView
+      {...props}
+      style={props?.style}
+      // @ts-expect-error
+      modifiers={props?.modifiers?.map((m) => m.__expo_shared_object_id__)}
+      type="pill"
+    />
+  ) as ShapeJSXElement;
 }
 
 function Circle(props: Pick<ShapeProps, 'radius' | 'verticesCount' | 'style' | 'color'>) {
-  return (<ShapeNativeView {...props} style={props?.style} type="circle" />) as ShapeJSXElement;
+  return (
+    <ShapeNativeView
+      {...props}
+      style={props?.style}
+      // @ts-expect-error
+      modifiers={props?.modifiers?.map((m) => m.__expo_shared_object_id__)}
+      type="circle"
+    />
+  ) as ShapeJSXElement;
 }
 
 function Rectangle(props: Pick<ShapeProps, 'smoothing' | 'cornerRounding' | 'style' | 'color'>) {
-  return (<ShapeNativeView {...props} style={props?.style} type="rectangle" />) as ShapeJSXElement;
+  return (
+    <ShapeNativeView
+      {...props}
+      style={props?.style}
+      // @ts-expect-error
+      modifiers={props?.modifiers?.map((m) => m.__expo_shared_object_id__)}
+      type="rectangle"
+    />
+  ) as ShapeJSXElement;
 }
 
 function Polygon(
@@ -71,7 +111,7 @@ function Polygon(
     <ShapeNativeView
       {...props}
       // @ts-expect-error
-      modifiers={props.modifiers?.map((m) => m.__expo_shared_object_id__)}
+      modifiers={props?.modifiers?.map((m) => m.__expo_shared_object_id__)}
       style={props?.style}
       type="polygon"
     />